backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.54.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
tenacity>=9.0.0

